Refactor: move IssueReader to service.command
